Develop

+++ src/main/java/com/nttdata/dockerized/postgresql/service/UserServiceImpl.java
@@ -1,5 +1,8 @@
 package com.nttdata.dockerized.postgresql.service;
 
+import com.nttdata.dockerized.postgresql.exception.BadRequestException;
+import com.nttdata.dockerized.postgresql.exception.InternalServerErrorException;
+import com.nttdata.dockerized.postgresql.exception.NotFoundException;
 import com.nttdata.dockerized.postgresql.model.entity.User;
 import com.nttdata.dockerized.postgresql.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,29 +23,45 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User findById(Long id) {
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Usuario no encontrado", "404"));
     }
 
     @Override
     public User save(User user) {
-        user.setActive(Boolean.TRUE);
-        return userRepository.save(user);
+        try {
+            if (user.getName() == null || user.getName().isBlank() ||
+                    user.getEmail() == null || user.getEmail().isBlank()){
+                throw new BadRequestException("El Email y el nombre son obligatorios");
+            }
+
+            user.setActive(Boolean.TRUE);
+            return userRepository.save(user);
+
+        } catch (BadRequestException e) {
+            throw e;
+        }catch (Exception e){
+            throw new InternalServerErrorException("Error interno al guardar el usuario: " + e.getMessage());
+        }
     }
 
     @Override
     public User updateById(Long id, User user) {
         return userRepository.findById(id)
-                .map(existing -> {
-                    existing.setName(user.getName());
-                    existing.setEmail(user.getEmail());
-                    existing.setActive(user.getActive());
-                    existing.setRegistrationDate(user.getRegistrationDate());
-                    return userRepository.save(existing);
-                }).orElse(null);
+                .map(userExisting -> {
+                    userExisting.setName(user.getName());
+                    userExisting.setEmail(user.getEmail());
+                    userExisting.setActive(user.getActive());
+                    userExisting.setRegistrationDate(user.getRegistrationDate());
+                    return userRepository.save(userExisting);
+                }).orElseThrow(() -> new NotFoundException("Usuario no encontrado para actualizar", "404"));
     }
 
     @Override
     public void deleteById(Long id) {
+        if (!userRepository.existsById(id)){
+            throw new NotFoundException("Usuario no encontrado para eliminar", "404");
+        }
         userRepository.deleteById(id);
     }
 }
